pyproject.toml
[project]
name = "proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "2captcha-python>=1.5.1",
    "aiohttp>=3.12.14",
    "async-cache>=1.1.1",
    "beautifulsoup4>=4.13.4",
    "cloudscraper>=1.2.71",
    "curl-cffi>=0.12.0",
    "fastapi>=0.116.1",
    "firebase-admin>=7.0.0",
    "pydantic-settings>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "redis>=6.2.0",
    "rich>=14.1.0",
    "selenium-base>=4.40.6",
    "uvicorn[standard]>=0.35.0",
]
